requirements.txt
flask
ngrok
torch
openai-whisper
